DAH-2334 reorganize
