+++ app/javascript/components/LotteryResultsPdfGenerator/components/DropTarget.js
