+++ app/javascript/components/LotteryResultsPdfGenerator/components/LotteryResults.js
